+++ src/config/IDIConfig.ts
@@ -1,5 +1,1 @@
-export interface IDIConfig {
-    pattern: string;
-    ignorePatterns: string[];
-    printLogo: boolean;
-}
+export interface IDIConfig {}