+++ src/service/STDIOUtils.ts
@@ -2,7 +2,6 @@ import { IServiceResponse } from './types/ServiceResponse';
 
 export class STDIOUtils {
 
-    //TODO use bytearray instead
     private static stdinBuffer = '';
 
     static read = (buffer: Buffer, onReadComplete: (data: string) => void): void => {