+++ src/service/handlers/ProcessFilesHandler.ts
@@ -2,8 +2,6 @@ import { ICommandHandler } from './ICommandHandler';
 import { IProcessFilesResponse } from '../types/process-files/IProcessFilesResponse';
 import { CompilationContext } from '../../compilation-context/CompilationContext';
 import { PathResolver } from '../../core/ts-helpers/path-resolver/PathResolver';
-import { FileSystem } from '../../file-system/FileSystem';
-import minimatch from 'minimatch';
 import { ContextRepository } from '../../core/context/ContextRepository';
 import { SourceFilesCache } from '../../core/ts-helpers/source-files-cache/SourceFilesCache';
 import { BeanRepository } from '../../core/bean/BeanRepository';
@@ -11,12 +9,12 @@ import { BeanDependenciesRepository } from '../../core/bean-dependencies/BeanDep
 import { DependencyGraph } from '../../core/connect-dependencies/DependencyGraph';
 import { LifecycleMethodsRepository } from '../../core/context-lifecycle/LifecycleMethodsRepository';
 import { PathResolverCache } from '../../core/ts-helpers/path-resolver/PathResolverCache';
-import { ConfigLoader } from '../../config/ConfigLoader';
 import { IDisposable } from '../types/IDisposable';
 import { IProcessFilesCommand } from '../types/process-files/IProcessFilesCommand';
 import { AbstractCompilationMessage } from '../../compilation-context/messages/AbstractCompilationMessage';
 import { StatisticsCollector } from '../statistics/StatisticsCollector';
-import { processContexts } from '../../core/build-context/processContexts';
+import { getTransformerFactory } from '../../core/transformers/getTransformerFactory';
+import ts from 'typescript';
 
 export class ProcessFilesHandler implements ICommandHandler<IProcessFilesCommand, Promise<IProcessFilesResponse>>, IDisposable {
     constructor(
@@ -26,13 +24,16 @@ export class ProcessFilesHandler implements ICommandHandler<IProcessFilesCommand
     async invoke(command: IProcessFilesCommand): Promise<IProcessFilesResponse> {
         try {
             const compilationContext = new CompilationContext();
+            const transformerFactory = getTransformerFactory(compilationContext);
             PathResolver.init();
 
-            const contextPaths = this.getContextPaths();
+            const sourceFiles = command.filesToProcess
+                .map(it => SourceFilesCache.getSourceFileByPath(it));
 
-            contextPaths.forEach(path => {
-                processContexts(compilationContext, SourceFilesCache.getSourceFileByPath(path));
-            });
+            ts.transform(
+                sourceFiles,
+                [transformerFactory],
+            );
 
             const affectedFiles = new Set<string>();
 
@@ -55,13 +56,6 @@ export class ProcessFilesHandler implements ICommandHandler<IProcessFilesCommand
         LifecycleMethodsRepository.clear();
         PathResolver.clear();
         PathResolverCache.clear();
-    }
-
-    private getContextPaths(): string[] {
-        const filePaths = FileSystem.getAllFilePaths();
-        const {pattern} = ConfigLoader.load();
-
-        return minimatch.match(filePaths, pattern);
     }
 
     private collectCompilationMessages(